+++ src/utils/index.js
@@ -1,1 +1,1 @@
-export * from './utils';
+export * from './render';
